improved CPU configuration handling

+++ src/virtual/vcpu.c
@@ -12,7 +12,7 @@ struct vcpu vcpus[DTB_NR_CPUS];
 uint64
 get_vhartid(struct vm_config *conf)
 {
-	int tmp = conf->cpu_affinity;
+	uint64 tmp = conf->cpu_affinity;
 	uint64 vhartid = 0;
 	for (uint64 i = 0; i < get_hartid(); i++) {
 		if (tmp & 1) {
